requirements.txt
connect-sdk == 20.1